add lucide react

+++ server/src/modules/Todo/TodoRepository.ts
@@ -3,9 +3,10 @@ import type { Result, Rows } from "../../../database/client";
 import type { NewTodo, Todo, UpdateTodo } from "../../types/todo";
 
 class TodoRepository {
-  async readAll() {
+  async readByUserId(userId: number) {
     const [rows] = await databaseClient.query<Rows>(
-      "SELECT * FROM todo ORDER BY created_at DESC",
+      "SELECT * FROM todo WHERE user_id = ? ORDER BY created_at DESC",
+      [userId],
     );
     return rows as Todo[];
   }
@@ -16,24 +17,24 @@ class TodoRepository {
     );
     return result.insertId;
   }
-  async delete(id: number) {
+  async delete(id: number, userId: number) {
     const [result] = await databaseClient.query<Result>(
-      "DELETE FROM todo WHERE id = ?",
-      [id],
+      "DELETE FROM todo WHERE id = ? AND user_id = ?",
+      [id, userId],
     );
     return result.affectedRows;
   }
-  async update(id: number, todo: UpdateTodo) {
+  async update(id: number, todo: UpdateTodo, userId: number) {
     const [result] = await databaseClient.query<Result>(
-      "UPDATE todo SET title = ?, description = ? WHERE id = ?",
-      [todo.title, todo.description, id],
+      "UPDATE todo SET title = ?, description = ? WHERE id = ? AND user_id = ?",
+      [todo.title, todo.description, id, userId],
     );
     return result.affectedRows;
   }
-  async updateCompleted(id: number, is_completed: boolean) {
+  async updateCompleted(id: number, is_completed: boolean, userId: number) {
     const [result] = await databaseClient.query<Result>(
-      "UPDATE todo SET is_completed = ? WHERE id = ?",
-      [is_completed, id],
+      "UPDATE todo SET is_completed = ? WHERE id = ? AND user_id = ?",
+      [is_completed, id, userId],
     );
     return result.affectedRows;
   }
